remove autogen step (bad for out-of-source builds)
running autogen.sh apart from generating the configure script it also
runs it inside the source dir with default arguments, which causes the
creation of header files in the source tree that define constants such
as JEMALLOC_PREFIX.

This, in turn, causes those headers to override any
preprocessor definitions in the out-of-source include directory.

This is not easily detectable because the generated header files are in
the gitignore file.